move metrics/metrics.py to main dir, remove metric dir
